Add files via upload
This is a basic template of the GUI, and sets the coding foundation to implement our individual commands to the code. The UI contains 4 tabs, of which incorporate the needs of userAdmin. There are plenty of comments in the code to explain the different classes and functions that the JavaFX library has and how its used to create a UI.